Added patch to declare the __builtin_getlinearva() properly.
With that patch, the __builtin_getlinearva() will be declared
within the .def include file. That include file is used at leas
twice. In one pass, it is used create an array leading to out of
bounds access when just an enum is declared as it was previously
the case.

The patch will be squashed in the next commit.